Move dom streaming helpers to core/dom/stream
